Finished cartridge memory refactor
