+++ gb/cart/mbc1.go
@@ -1,5 +1,6 @@
 package cart
 
+// NewMBC1 returns a new MBC1 memory controller.
 func NewMBC1(data []byte) BankingController {
 	return &MBC1{
 		rom:     data,
@@ -8,7 +9,7 @@ func NewMBC1(data []byte) BankingController {
 	}
 }
 
-// ROM is a basic Gameboy cartridge.
+// MBC1 is a GameBoy cartridge that supports rom and ram banking.
 type MBC1 struct {
 	rom     []byte
 	romBank uint32
@@ -20,7 +21,7 @@ type MBC1 struct {
 	romBanking bool
 }
 
-// Read returns a value at a memory address in the ROM.
+// Read returns a value at a memory address in the ROM or RAM.
 func (r *MBC1) Read(address uint16) byte {
 	switch {
 	case address < 0x4000:
@@ -32,7 +33,7 @@ func (r *MBC1) Read(address uint16) byte {
 	}
 }
 
-// Write is not supported on a ROM cart.
+// WriteROM attempts to switch the ROM or RAM bank.
 func (r *MBC1) WriteROM(address uint16, value byte) {
 	switch {
 	case address < 0x2000:
@@ -69,9 +70,21 @@ func (r *MBC1) WriteROM(address uint16, value byte) {
 	}
 }
 
+// WriteRAM writes data to the ram if it is enabled.
 func (r *MBC1) WriteRAM(address uint16, value byte) {
 	if r.ramEnabled {
 		r.ram[(0x2000*r.ramBank)+uint32(address-0xA000)] = value
 	}
-	// TODO: what do if disabled
+}
+
+// GetSaveData returns the save data for this banking controller.
+func (r *MBC1) GetSaveData() []byte {
+	data := make([]byte, len(r.ram))
+	copy(data, r.ram)
+	return data
+}
+
+// LoadSaveData loads the save data into the cartridge.
+func (r *MBC1) LoadSaveData(data []byte) {
+	r.ram = data
 }
